requirements.txt
python-dotenv
requests
pytest
